+++ src/main/java/subway/config/handler/SubwayConstraintHandler.java
@@ -1,4 +1,4 @@
-package subway.config;
+package subway.config.handler;
 
 public class SubwayConstraintHandler {
 }
